initial merging of maps
